media-gfx/blender-bin: update ebuilds
Reorder variables to make pkgcheck happy.
Restrict test as there are none.
Introduce USE for cuda, hip and oneapi to remove libraries that link to not
installed libraries.
Update RDEPEND.
Wrap downloading in src_unpack in some validation logic.
Move code modification to src_prepare.
Add empty src_configure and src_compile so the default versions are not used.
Use ED instead of D in src_install, so this can work in prefix.

Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>